pyproject.toml
[project]
name = "gisual-interview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"alembic==1.15.2",
	"annotated-types==0.7.0",
	"anyio==4.9.0",
	"astroid==3.3.9",
	"asyncpg==0.30.0",
	"autopep8==2.0.4",
	"bcrypt==4.0.1",
	"certifi==2025.1.31",
	"cffi==1.17.1",
	"charset-normalizer==3.4.1",
	"click==8.1.8",
	"cryptography==44.0.2",
	"dill==0.3.9",
	"dnspython==2.7.0",
	"docstring-to-markdown==0.16",
	"email-validator==2.2.0",
	"fastapi==0.115.12",
	"fastapi-cli==0.0.7",
	"flake8==7.1.2",
	"geographiclib==2.0",
	"geopandas==1.0.1",
	"geopy==2.4.1",
	"greenlet==3.1.1",
	"h11==0.14.0",
	"httpcore==1.0.7",
	"httptools==0.6.4",
	"httpx==0.28.1",
	"idna==3.10",
	"importlib-metadata==8.6.1",
	"isort==6.0.1",
	"itsdangerous==2.2.0",
	"jedi==0.19.2",
	"jinja2==3.1.6",
	"joblib==1.4.2",
	"mako==1.3.9",
	"markdown-it-py==3.0.0",
	"markupsafe==3.0.2",
	"mccabe==0.7.0",
	"mdurl==0.1.2",
	"nodeenv==1.9.1",
	"numpy==2.2.4",
	"orjson==3.10.16",
	"packaging==24.2",
	"pandas==2.2.3",
	"parso==0.8.4",
	"passlib==1.7.4",
	"platformdirs==4.3.7",
	"pluggy==1.5.0",
	"psycopg2-binary==2.9.10",
	"pycodestyle==2.12.1",
	"pycparser==2.22",
	"pydantic==2.11.2",
	"pydantic-core==2.33.1",
	"pydantic-extra-types==2.10.3",
	"pydantic-settings==2.8.1",
	"pydocstyle==6.3.0",
	"pyflakes==3.2.0",
	"pygments==2.19.1",
	"pyjwt==2.10.1",
	"pylint==3.3.6",
	"pyogrio==0.10.0",
	"pyproj==3.7.1",
	"pyright==1.1.398",
	"python-dateutil==2.9.0.post0",
	"python-dotenv==1.1.0",
	"python-lsp-jsonrpc==1.1.2",
	"python-lsp-server==1.12.2",
	"python-multipart==0.0.20",
	"pytoolconfig==1.3.1",
	"pytz==2025.2",
	"pyyaml==6.0.2",
	"redis==5.2.1",
	"requests==2.32.3",
	"rich==14.0.0",
	"rich-toolkit==0.14.1",
	"rope==1.13.0",
	"ruff==0.11.4",
	"scikit-learn==1.6.1",
	"scipy==1.15.2",
	"shapely==2.1.0",
	"shellingham==1.5.4",
	"six==1.17.0",
	"sniffio==1.3.1",
	"snowballstemmer==2.2.0",
	"sqlalchemy==2.0.40",
	"sqlmodel==0.0.24",
	"starlette==0.46.1",
	"threadpoolctl==3.6.0",
	"tomlkit==0.13.2",
	"typer==0.15.2",
	"typing-extensions==4.13.1",
	"typing-inspection==0.4.0",
	"tzdata==2025.2",
	"ujson==5.10.0",
	"urllib3==2.3.0",
	"uvicorn==0.34.0",
	"uvloop==0.21.0",
	"watchfiles==1.0.4",
	"websockets==15.0.1",
	"whatthepatch==1.0.7",
	"yapf==0.43.0",
	"zipp==3.21.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
